backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
bcrypt>=4.1.0
PyJWT>=2.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

